#26 Rename time conversion module name from utils
Use a better name `edtf.convert` instead of
`edtf.utils`
