+++ edtf/__init__.py
@@ -1,5 +1,5 @@
 from edtf.parser.grammar import parse_edtf
 from edtf.natlang import text_to_edtf
 from edtf.parser.parser_classes import *
-from edtf.utils import dt_to_struct_time, struct_time_to_date, \
+from edtf.convert import dt_to_struct_time, struct_time_to_date, \
     struct_time_to_datetime, trim_struct_time
