+++ edtf/tests.py
@@ -3,14 +3,14 @@ import unittest
 from time import struct_time
 from datetime import datetime, date
 
-from edtf import utils
+from edtf import convert
 
 
 class TestConversions(unittest.TestCase):
 
     def test_dt_to_struct_time_for_datetime(self):
         now = datetime.now()
-        st = utils.dt_to_struct_time(now)
+        st = convert.dt_to_struct_time(now)
         # Check equal year, month, day, hours, minutes, seconds
         self.assertEqual(st[:6], now.timetuple()[:6])
         # Confirm 'extra' fields are set to defaults
@@ -18,7 +18,7 @@ class TestConversions(unittest.TestCase):
 
     def test_dt_to_struct_time_for_date(self):
         today = date.today()
-        st = utils.dt_to_struct_time(today)
+        st = convert.dt_to_struct_time(today)
         # Check equal year, month, day
         self.assertEqual(st[:3], today.timetuple()[:3])
         # Confirm time fields are zeroed
@@ -28,15 +28,15 @@ class TestConversions(unittest.TestCase):
 
     def test_struct_time_to_date(self):
         st = struct_time(
-            [2018, 4, 19] + utils.TIME_EMPTY_TIME + utils.TIME_EMPTY_EXTRAS)
+            [2018, 4, 19] + convert.TIME_EMPTY_TIME + convert.TIME_EMPTY_EXTRAS)
         d = date(*st[:3])
-        self.assertEqual(d, utils.struct_time_to_date(st))
+        self.assertEqual(d, convert.struct_time_to_date(st))
 
     def test_struct_time_to_datetime(self):
         st = struct_time(
-            [2018, 4, 19] + [10, 13, 54] + utils.TIME_EMPTY_EXTRAS)
+            [2018, 4, 19] + [10, 13, 54] + convert.TIME_EMPTY_EXTRAS)
         dt = datetime(*st[:6])
-        converted_dt = utils.struct_time_to_datetime(st)
+        converted_dt = convert.struct_time_to_datetime(st)
         self.assertEqual(dt, converted_dt)
         # Note that 'extra' fields are auto-populated by `datetime` module
         self.assertEqual(converted_dt.timetuple()[6:], (3, 109, -1))
@@ -44,7 +44,7 @@ class TestConversions(unittest.TestCase):
     def test_trim_struct_time(self):
         now = datetime.now()
         st = now.timetuple()
-        trimmed_st = utils.trim_struct_time(st)
+        trimmed_st = convert.trim_struct_time(st)
         # Confirm trimmed `struct_time` has expected date/time values
         self.assertEqual(
             trimmed_st[:6],
